# optimizer: improve `lift_comparison_leaves!` performance