https://github.com/JuliaLang/julia/pull/43227 seems to have introduced
some allocation increase:
> before #43227 (2682819f06a6344e590efa9aca506e32ca07700b)
```
~/julia/packages/EscapeAnalysis master
❯ ~/julia/julia5/usr/bin/julia -e 'using EscapeAnalysis; code_escapes(sin, (Int,)); @time code_escapes(Base.init_stdio, (Ptr{Cvoid},)); @time code_escapes(println, (QuoteNode,))'
  4.693671 seconds (24.01 M allocations: 1.261 GiB, 20.25% gc time, 90.90% compilation time)
  0.377396 seconds (2.36 M allocations: 128.876 MiB, 14.03% gc time, 99.98% compilation time)
```

> #43227 (cdd2f623eb03de024b8437fb9b0a896968202b21)
```
~/julia/packages/EscapeAnalysis master
❯ jld -e 'using EscapeAnalysis; code_escapes(sin, (Int,)); @time code_escapes(Base.init_stdio, (Ptr{Cvoid},)); @time code_escapes(println, (QuoteNode,))'
  4.93641 seconds (24.67 M allocations: 1.275 GiB, 17.72% gc time, 89.63% compilation time)
  0.406880 seconds (2.42 M allocations: 130.121 MiB, 13.46% gc time, 99.98% compilation time)
```

This commit reduces memory allocations a bit:
> this commit
```
~/julia/packages/EscapeAnalysis master 9s
❯ jld -e 'using EscapeAnalysis; code_escapes(sin, (Int,)); @time code_escapes(Base.init_stdio, (Ptr{Cvoid},)); @time code_escapes(println, (QuoteNode,))'
  4.474278 seconds (24.10 M allocations: 1.263 GiB, 19.80% gc time, 92.11% compilation time)
  0.455346 seconds (2.37 M allocations: 128.992 MiB, 15.22% gc time, 99.98% compilation time)
```