@aviatesk 2022-01-26
Hmm, very weird. After rebasing against master and taking in #43922, it no longer reduces the allocations if we add `@inline` to `lift_comparison_leaves!`..

I tried various minor optimizations but failed to reduce allocations. Currently I get the following results:

> baseline -- before #43227 (2682819f06a6344e590efa9aca506e32ca07700b)
```
~/julia/julia5 master ⇣ 10s
❯ ./usr/bin/julia ../CCProfiler.jl
  1.846204 seconds (5.83 M allocations: 310.795 MiB, 9.86% gc time, 70.37% compilation time)
  0.722648 seconds (5.26 M allocations: 281.215 MiB, 26.92% gc time, 100.00% compilation time)
  0.707713 seconds (5.26 M allocations: 281.215 MiB, 25.43% gc time, 100.00% compilation time)
  0.049564 seconds (449.05 k allocations: 24.850 MiB, 23.48% gc time, 99.92% compilation time)
  3.218099 seconds (21.66 M allocations: 1.117 GiB, 19.51% gc time, 100.00% compilation time)
  0.566937 seconds (4.70 M allocations: 250.310 MiB, 21.09% gc time, 99.99% compilation time)
  0.172283 seconds (1.50 M allocations: 81.192 MiB, 99.98% compilation time)
```

> master (3f0ae6e942c22b64742a185fcf05b8e473df83e9)
```
~/julia/julia4 master 13s
❯ ./usr/bin/julia ../CCProfiler.jl
2.215941 seconds (5.97 M allocations: 313.993 MiB, 10.33% gc time, 68.60% compilation time)
0.839248 seconds (5.39 M allocations: 284.056 MiB, 24.14% gc time, 100.00% compilation time)
0.852634 seconds (5.39 M allocations: 284.056 MiB, 24.94% gc time, 100.00% compilation time)
0.057837 seconds (458.27 k allocations: 25.066 MiB, 20.75% gc time, 99.92% compilation time)
4.654196 seconds (22.21 M allocations: 1.129 GiB, 17.84% gc time, 100.00% compilation time)
0.578950 seconds (4.81 M allocations: 253.040 MiB, 9.42% gc time, 99.99% compilation time)
0.306086 seconds (1.53 M allocations: 81.935 MiB, 47.38% gc time, 99.99% compilation time)
```

> this commit (571cd4d)
```
~/julia/julia avi/perf 9s
❯ ./usr/bin/julia ../CCProfiler.jl
  2.170536 seconds (5.97 M allocations: 313.798 MiB, 13.72% gc time, 70.37% compilation time)
  0.908299 seconds (5.39 M allocations: 283.866 MiB, 26.01% gc time, 100.00% compilation time)
  0.844291 seconds (5.39 M allocations: 283.866 MiB, 23.86% gc time, 100.00% compilation time)
  0.123409 seconds (457.79 k allocations: 25.032 MiB, 58.01% gc time, 99.97% compilation time)
  4.125487 seconds (22.19 M allocations: 1.128 GiB, 19.41% gc time, 100.00% compilation time)
  0.604038 seconds (4.81 M allocations: 252.872 MiB, 19.23% gc time, 99.99% compilation time)
  0.186763 seconds (1.53 M allocations: 81.850 MiB, 21.64% gc time, 99.98% compilation time)
```

where, `CCProfiler.jl` is defined as like:
```julia
# setup new AbstractInterpreter
const CC = Core.Compiler
import .CC: MethodInstance, CodeInstance, WorldRange, WorldView

struct CCProfilerCache
    dict::IdDict{MethodInstance,CodeInstance}
end
struct CCProfiler <: CC.AbstractInterpreter
    interp::CC.NativeInterpreter
    cache::CCProfilerCache
    CCProfiler(world = Base.get_world_counter();
        interp = CC.NativeInterpreter(world),
        cache = CCProfilerCache(IdDict{MethodInstance,CodeInstance}())
        ) = new(interp, cache)
end
CC.InferenceParams(profiler::CCProfiler) = CC.InferenceParams(profiler.interp)
CC.OptimizationParams(profiler::CCProfiler) = CC.OptimizationParams(profiler.interp)
CC.get_world_counter(profiler::CCProfiler) = CC.get_world_counter(profiler.interp)
CC.get_inference_cache(profiler::CCProfiler) = CC.get_inference_cache(profiler.interp)
CC.code_cache(profiler::CCProfiler) = WorldView(profiler.cache, WorldRange(CC.get_world_counter(profiler)))
CC.get(wvc::WorldView{<:CCProfilerCache}, mi::MethodInstance, default) = get(wvc.cache.dict, mi, default)
CC.getindex(wvc::WorldView{<:CCProfilerCache}, mi::MethodInstance) = getindex(wvc.cache.dict, mi)
CC.haskey(wvc::WorldView{<:CCProfilerCache}, mi::MethodInstance) = haskey(wvc.cache.dict, mi)
CC.setindex!(wvc::WorldView{<:CCProfilerCache}, ci::CodeInstance, mi::MethodInstance) = setindex!(wvc.cache.dict, ci, mi)

# compile things, and check that the global cache isn't shared across inferences
@time Base.return_types(println, (QuoteNode,), CCProfiler());
@time Base.return_types(println, (QuoteNode,), CCProfiler());
@time Base.return_types(println, (QuoteNode,), CCProfiler());

@time Base.return_types(sin, (Int,), CCProfiler());
@time Base.return_types(Base.init_stdio, (Ptr{Cvoid},), CCProfiler());
@time Base.return_types(CC.construct_ssa!, (Core.CodeInfo,CC.IRCode,CC.DomTree,Vector{CC.SlotInfo},Vector{Any}), CCProfiler())
@time Base.return_types(CC.domsort_ssa!, (CC.IRCode,CC.DomTree), CCProfiler())
```

So the allocations may be inevitable if we call `lift_comparison!` for those three builtin calls (, which may happen somewhat often).